+++ src/main/java/me/zerosquare/simplemodel/Model.java
@@ -1,5 +1,6 @@
 package me.zerosquare.simplemodel;
 
+import me.zerosquare.simplemodel.annotations.Table;
 import me.zerosquare.simplemodel.internal.Connector;
 import me.zerosquare.simplemodel.internal.Logger;
 import org.apache.commons.lang3.StringUtils;
