+++ src/main/java/me/zerosquare/simplemodel/ModelData.java
@@ -1,6 +1,7 @@
 package me.zerosquare.simplemodel;
 
 import me.zerosquare.simplemodel.Model.QueryType;
+import me.zerosquare.simplemodel.annotations.Column;
 import me.zerosquare.simplemodel.internal.Logger;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.tuple.MutablePair;
