+++ src/main/java/me/zerosquare/simplemodel/annotations/Column.java
@@ -1,13 +1,12 @@
-package me.zerosquare.simplemodel;
+package me.zerosquare.simplemodel.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target(ElementType.TYPE)
+@Target(ElementType.FIELD)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface Table {
-  String name();
+public @interface Column {
+  String name() default "";
 }
-
