+++ src/main/java/me/zerosquare/simplemodel/annotations/Table.java
@@ -1,12 +1,13 @@
-package me.zerosquare.simplemodel;
+package me.zerosquare.simplemodel.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target(ElementType.FIELD)
+@Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface Column {
-  String name() default "";
+public @interface Table {
+  String name();
 }
+
